+++ www/i18n/de.js
@@ -919,6 +919,7 @@ defLang('de', (function(){
       'post.save'             : "Sichern",
       'post.edited'           : "Bearbeitet",
       'post.pv'               : "Privat",
+      'post.unsent'               : "Noch nicht gesendet",
       'post.pv.on'            : "Nur du siehst das",
       'post.pv.off'           : "Alle sehen das",
       'post.draft.save'       : "Entwurf sichern",