+++ www/i18n/en.js
@@ -844,6 +844,7 @@ defLang('en', (function(){
       'post.save'             : "Save",
       'post.edited'           : "Edited",
       'post.pv'               : "Private",
+      'post.unsent'               : "Not sent yet",
       'post.pv.on'            : "Only you will see this",
       'post.pv.off'           : "Everybody will see this",
       'post.draft.save'       : "Save a draft",