+++ www/i18n/it.js
@@ -932,6 +932,7 @@ defLang('it', (function(){
       'post.save'             : "Salva",
       'post.edited'           : "Modificato",
       'post.pv'               : "Privato",
+      'post.unsent'               : "Non ancora inviato",
       'post.pv.on'            : "Lo vedrai solo tu",
       'post.pv.off'           : "Lo vedranno tutti",
       'post.draft.save'       : "Salvare una bozza",