+++ www/i18n/ko.js
@@ -1105,6 +1105,7 @@ defLang('ko', (function(){
       'post.save'             : "저장",
       'post.edited'           : "편집됨",
       'post.pv'               : "나만 보기",
+      'post.unsent'               : "아직 보내지 못했습니다",
       'post.pv.on'            : "나만 볼 수 있습니다",
       'post.pv.off'           : "모두가 볼 수 있습니다",
       'post.draft.save'       : "임시저장",