+++ www/i18n/pt.js
@@ -937,6 +937,7 @@ defLang('pt', (function(){
       'post.save'             : "Guardar",
       'post.edited'           : "Editado",
       'post.pv'               : "Privado",
+      'post.unsent'               : "Ainda não enviado",
       'post.pv.on'            : "Só você vai ver",
       'post.pv.off'           : "Todos vão ver",
       'post.draft.save'       : "Guardar rascunho",